ChatWithExtractedText/requirements.txt
streamlit==1.10.0
google-generativeai
Pillow
pytesseract
pandas